New fields and flexible label-column matching with regexps.
Programmable web made slight changes to labels shown in API description changes that break the current field assignment. To avoid constant updates on the fields array, regexps are generated from that array in order to match the labels used on the scrapped html.

New fields used on descriptions were added as well, no field was renamed in the database so queries written for the database in previous version should still be compatible. The only breaking change on ProgrammableWeb side was 'API Homepage' changed to 'Api Portal Home Page'. Together with the aforementioned regexps, a 'Home_Page' -> 'Homepage' rule was added to keep things working.